move filter classes into their own file.
